+++ buttons/buttons_search_quicksearch.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//15/11/24
+//25/11/24
 
 /*
 	Quicksearch for same....
@@ -92,22 +92,15 @@ var newButtonsProperties = { // NOSONAR[global]
 		},
 		{
 			name: 'Same Artist(s)',
-			query: [
-				'(ALBUM ARTIST PRESENT AND ',
-				'(ALBUM ARTIST IS #ALBUM ARTIST#)',
-				') OR (',
-				'ALBUM ARTIST MISSING AND ',
-				'(ARTIST IS #ARTIST#)',
-				')'
-			]
+			query: globTags.artist + ' IS #' + globTags.artistRaw + '#'
 		},
 		{
 			name: 'Same Title and Artist(s)',
-			query: globQuery.compareTitle + ' AND (ARTIST IS #ARTIST#)'
+			query: globQuery.compareTitle + ' AND (' + globTags.artist + ' IS #' + globTags.artistRaw + '#)'
 		},
 		{
 			name: 'Same Title, Artist(s) & Date',
-			query: globQuery.compareTitle + ' AND (ARTIST IS #ARTIST#) AND (' + _q(globTags.date) + ' IS #' + globTags.date + '#)'
+			query: globQuery.compareTitle + ' AND (' + globTags.artist + ' IS #' + globTags.artistRaw + '#) AND (' + _q(globTags.date) + ' IS #' + globTags.date + '#)'
 		}
 	]), { func: isJSON }],
 	sortTF: ['Sorting TF expression', globTags.artist + '|%ALBUM%|%TRACK%', { func: isStringWeak }, globTags.artist + '|%ALBUM%|%TRACK%'],
@@ -211,6 +204,12 @@ function quickSearchMenu({ bSimulate = false } = {}) {
 		this.selItems = plman.ActivePlaylist !== -1 ? plman.GetPlaylistSelectedItems(plman.ActivePlaylist) : null;
 		if (!this.selItems || !this.selItems.Count) { this.selItems = null; console.log('Quicksearch: No selected items.'); }
 	}
+	const multiTags = ['artist', 'genre', 'style', globTags.artistRaw, globTags.genre, globTags.style]
+		.map((t) => t.toLowerCase());
+	const notMultTags = (e) => {
+		const query = (e.query || '').toLowerCase();
+		return !query || !multiTags.some((t) => query.includes(t));
+	};
 	const flags = this.selItems && this.selItems.Count ? MF_STRING : MF_GRAYED;
 	// Entry list
 	const queryFilter = JSON.parse(this.buttonsProperties.entries[1]);
@@ -234,7 +233,7 @@ function quickSearchMenu({ bSimulate = false } = {}) {
 				menu.newEntry({
 					entryText: queryObj.name, func: () => {
 						let query = queryObj.query;
-						if ((query.indexOf('#') !== -1 || (Array.isArray(query) && query.some((q) => q.indexOf('#') !== -1))) && !fb.GetFocusItem(true)) { fb.ShowPopupMessage('Can not evaluate query without a selection:\n' + queryObj.query, 'Quicksearch'); return; }
+						if ((query.includes('#') || (Array.isArray(query) && query.some((q) => q.includes('#')))) && !fb.GetFocusItem(true)) { fb.ShowPopupMessage('Can not evaluate query without a selection:\n' + queryObj.query, 'Quicksearch'); return; }
 						const bShift = utils.IsKeyPressed(VK_SHIFT);
 						const bCtrl = utils.IsKeyPressed(VK_CONTROL);
 						if (this.buttonsProperties.bEvalSel[1]) {
@@ -273,7 +272,7 @@ function quickSearchMenu({ bSimulate = false } = {}) {
 				let query = '';
 				try { query = utils.InputBox(window.ID, 'Enter query:\nAlso allowed dynamic variables, like #ARTIST#, which will be replaced with ' + (this.buttonsProperties.bEvalSel[1] ? 'selected items\' values.' : 'focused item\'s value.') + '\n\nPressing Shift while clicking on \'OK\' will open the search window.\nPressing Ctrl will create an AutoPlaylist.', 'Quicksearch', this.buttonsProperties.lastQuery[1] || 'TITLE IS #TITLE#', true); }
 				catch (e) { return; }
-				if (query.indexOf('#') !== -1 && !fb.GetFocusItem(true)) { fb.ShowPopupMessage('Can not evaluate query without a selection:\n' + query, 'Quicksearch'); return; }
+				if (query.includes('#') && !fb.GetFocusItem(true)) { fb.ShowPopupMessage('Can not evaluate query without a selection:\n' + query, 'Quicksearch'); return; }
 				if (!query.length) { return; }
 				// Playlist
 				const bShift = utils.IsKeyPressed(VK_SHIFT);
@@ -302,7 +301,7 @@ function quickSearchMenu({ bSimulate = false } = {}) {
 	}
 	menu.newSeparator();
 	{	// Begin with
-		const beginMenu = menu.newMenu('Begins with...');
+		const beginMenu = menu.newMenu('Begins with');
 		menu.newEntry({ menuName: beginMenu, entryText: 'Simulates \'%TAG% IS VALUE*\':', flags: MF_GRAYED });
 		menu.newSeparator(beginMenu);
 		[
@@ -319,11 +318,11 @@ function quickSearchMenu({ bSimulate = false } = {}) {
 				queryObj.name = queryObj.name.length > 40 ? queryObj.name.substring(0, 40) + ' ...' : queryObj.name;
 				// Entries
 				if (Array.isArray(queryObj.query)) { return; }
-				if (queryObj.query.count('#') === 2 && queryObj.query.indexOf('$') === -1) {
+				if (queryObj.query.count('#') === 2 && !queryObj.query.includes('$')) {
 					menu.newEntry({
 						menuName: beginMenu, entryText: queryObj.name, func: () => {
-							let query = queryObj.query + '*';
-							if (query.indexOf('#') !== -1 && !fb.GetFocusItem(true)) { fb.ShowPopupMessage('Can not evaluate query without a selection:\n' + queryObj.query, 'Quicksearch'); return; }
+							let query = queryObj.query.replace(/#$/, '*#');
+							if (query.includes('#') && !fb.GetFocusItem(true)) { fb.ShowPopupMessage('Can not evaluate query without a selection:\n' + queryObj.query, 'Quicksearch'); return; }
 							const bShift = utils.IsKeyPressed(VK_SHIFT);
 							const bCtrl = utils.IsKeyPressed(VK_CONTROL);
 							if (this.buttonsProperties.bEvalSel[1]) {
@@ -356,7 +355,7 @@ function quickSearchMenu({ bSimulate = false } = {}) {
 		});
 	}
 	{	// Includes
-		const beginMenu = menu.newMenu('Partially includes...');
+		const beginMenu = menu.newMenu('Partially includes');
 		menu.newEntry({ menuName: beginMenu, entryText: 'Simulates \'%TAG% HAS VALUE\':', flags: MF_GRAYED });
 		menu.newSeparator(beginMenu);
 		[
@@ -373,11 +372,11 @@ function quickSearchMenu({ bSimulate = false } = {}) {
 				queryObj.name = queryObj.name.length > 40 ? queryObj.name.substring(0, 40) + ' ...' : queryObj.name;
 				// Entries
 				if (Array.isArray(queryObj.query)) { return; }
-				if (queryObj.query.count('#') === 2 && queryObj.query.indexOf('$') === -1) {
+				if (queryObj.query.count('#') === 2 && !queryObj.query.includes('$')) {
 					menu.newEntry({
 						menuName: beginMenu, entryText: queryObj.name, func: () => {
-							let query = queryObj.query;
-							if (query.indexOf('#') !== -1 && !fb.GetFocusItem(true)) { fb.ShowPopupMessage('Can not evaluate query without a selection:\n' + queryObj.query, 'Quicksearch'); return; }
+							let query = queryObj.query.replaceAll(' IS ', ' HAS ');
+							if (query.includes('#') && !fb.GetFocusItem(true)) { fb.ShowPopupMessage('Can not evaluate query without a selection:\n' + queryObj.query, 'Quicksearch'); return; }
 							const bShift = utils.IsKeyPressed(VK_SHIFT);
 							const bCtrl = utils.IsKeyPressed(VK_CONTROL);
 							if (this.buttonsProperties.bEvalSel[1]) {
@@ -410,26 +409,29 @@ function quickSearchMenu({ bSimulate = false } = {}) {
 		});
 	}
 	{	// Partial
-		const partialMenu = menu.newMenu('Partial match...');
+		const partialMenu = menu.newMenu('Partial match');
 		menu.newEntry({ menuName: partialMenu, entryText: 'Matches any value partially equal:', flags: MF_GRAYED });
 		menu.newSeparator(partialMenu);
 		// Mutate original queries into partial matches
 		[
 			...queryFilter,
-			{ name: 'By Title and same Artist', query: globQuery.compareTitle, queryPost: ' AND (ARTIST IS #ARTIST#)' }
-		].map((queryObj) => {
+			{ name: 'By Title and same Artist', query: globQuery.compareTitle, queryPost: ' AND (' + globTags.artist + ' IS #' + globTags.artistRaw + '#)' }
+		].filter(notMultTags).map((queryObj) => {
 			if (Array.isArray(queryObj.query)) { return void (0); }
 			if (Object.hasOwn(queryObj, 'query') && queryObj.query.count('#') === 2) {
 				const dynTF = queryObj.query.match(/#.*#/)[0];
 				if (dynTF && dynTF.length) {
-					const bIsFunc = dynTF.indexOf('$') !== -1;
-					const statTF = bIsFunc ? dynTF.replaceAll('#', '') : '%' + dynTF.replaceAll('#', '') + '%';
+					const bIsFunc = dynTF.includes('$');
+					const statTF = bIsFunc
+						? dynTF.replaceAll('#', '')
+						: '%' + dynTF.replaceAll('#', '') + '%';
 					return {
 						name: queryObj.name.replace('Same', 'By'),
 						// Reduce length in 1 for dates, so it matches the same decade!
-						query: (/DATE/i.test(queryObj.query) !== true
-							? '"$puts(val,' + dynTF + ')$puts(vallen,$len($get(val)))$puts(min,$min($get(vallen),$if2($strchr($get(val),\'(\'),$get(vallen)),$if2($strchr($get(val),\'[\'),$get(vallen)),$len(' + statTF + ')))$stricmp($left($get(val),$get(min)),$left(' + statTF + ',$get(min)))" IS 1'
-							: '"$puts(val,' + dynTF + ')$puts(min,$sub($min($get(vallen),$len(' + statTF + ')),1))$stricmp($left($get(val),$get(min)),$left(' + statTF + ',$get(min)))" IS 1'
+						query: (
+							/DATE/i.test(queryObj.query)
+								? '"$puts(val,' + dynTF + ')$puts(vallen,$len($get(val)))$puts(min,$sub($min($get(vallen),$len(' + statTF + ')),1))$stricmp($left($get(val),$get(min)),$left(' + statTF + ',$get(min)))" IS 1'
+								: '"$puts(val,' + dynTF + ')$puts(vallen,$len($get(val)))$puts(min,$min($get(vallen),$if2($strchr($get(val),\'(\'),$get(vallen)),$if2($strchr($get(val),\'[\'),$get(vallen)),$len(' + statTF + ')))$stricmp($left($get(val),$get(min)),$left(' + statTF + ',$get(min)))" IS 1'
 						) + (queryObj.queryPost || '')
 					};
 				}
@@ -450,7 +452,96 @@ function quickSearchMenu({ bSimulate = false } = {}) {
 				menu.newEntry({
 					menuName: partialMenu, entryText: queryObj.name, func: () => {
 						let query = queryObj.query;
-						if (query.indexOf('#') !== -1 && !fb.GetFocusItem(true)) { fb.ShowPopupMessage('Can not evaluate query without a selection:\n' + queryObj.query, 'Quicksearch'); return; }
+						if (query.includes('#') && !fb.GetFocusItem(true)) { fb.ShowPopupMessage('Can not evaluate query without a selection:\n' + queryObj.query, 'Quicksearch'); return; }
+						const bShift = utils.IsKeyPressed(VK_SHIFT);
+						const bCtrl = utils.IsKeyPressed(VK_CONTROL);
+						if (this.buttonsProperties.bEvalSel[1]) {
+							if (bShift || bCtrl) {
+								query = dynamicQueryProcess({ query, handleList: this.selItems });
+								if (query) {
+									if (bShift && !bCtrl) { fb.ShowLibrarySearchUI(query); }
+									else if (!bShift && bCtrl) { plman.CreateAutoPlaylist(plman.PlaylistCount, playlistName, query); }
+									else { dynamicQuery({ query, sort: (bOmitSortPls ? null : queryObj.sort || { tfo: sortTF }), handleList: this.selItems, playlistName, source: plman.GetPlaylistItems(plman.ActivePlaylist) }); }
+								}
+							} else {
+								dynamicQuery({ query, sort: queryObj.sort || { tfo: sortTF }, handleList: this.selItems, playlistName });
+							}
+						} else {
+							if (bShift || bCtrl) { // NOSONAR
+								query = dynamicQueryProcess({ query });
+								if (query) {
+									if (bShift && !bCtrl) { fb.ShowLibrarySearchUI(query); }
+									else if (!bShift && bCtrl) { plman.CreateAutoPlaylist(plman.PlaylistCount, playlistName, query); }
+									else { dynamicQuery({ query, sort: (bOmitSortPls ? null : queryObj.sort || { tfo: sortTF }), handleList: this.selItems, playlistName, source: plman.GetPlaylistItems(plman.ActivePlaylist) }); }
+								}
+							} else {
+								dynamicQuery({ query, sort: queryObj.sort || { tfo: sortTF }, playlistName });
+							}
+						}
+					}, flags, data: { bDynamicMenu: true }
+				});
+			}
+		});
+	}
+	{	// Similar
+		const partialMenu = menu.newMenu('Similar match');
+		menu.newEntry({ menuName: partialMenu, entryText: 'Matches any similar value:', flags: MF_GRAYED });
+		menu.newSeparator(partialMenu);
+		// Mutate original queries into partial matches
+		[
+			...queryFilter,
+			{ name: 'By Title and same Artist', query: globQuery.compareTitle, queryPost: ' AND (' + globTags.artist + ' IS #' + globTags.artistRaw + '#)' }
+		].map((queryObj) => {
+			if (Array.isArray(queryObj.query)) { return void (0); }
+			if (Object.hasOwn(queryObj, 'query') && queryObj.query.count('#') === 2) {
+				const dynTF = queryObj.query.match(/#.*#/)[0];
+				if (dynTF && dynTF.length) {
+					const bIsFunc = dynTF.includes('$');
+					const statTF = bIsFunc
+						? dynTF.replaceAll('#', '')
+						: '%' + dynTF.replaceAll('#', '') + '%';
+					return {
+						name: queryObj.name.replace('Same', 'By'),
+						// Reduce length in 1 for dates, so it matches the same decade!
+						query: [
+							...(notMultTags(queryObj)
+								? [
+									'(',
+									(
+										/DATE/i.test(queryObj.query)
+											? '"$puts(val,' + dynTF + ')$puts(vallen,$len($get(val)))$puts(min,$sub($min($get(vallen),$len(' + statTF + ')),1))$stricmp($left($get(val),$get(min)),$left(' + statTF + ',$get(min)))" IS 1'
+											: '"$puts(val,' + dynTF + ')$puts(vallen,$len($get(val)))$puts(min,$min($get(vallen),$if2($strchr($get(val),\'(\'),$get(vallen)),$if2($strchr($get(val),\'[\'),$get(vallen)),$len(' + statTF + ')))$stricmp($left($get(val),$get(min)),$left(' + statTF + ',$get(min)))" IS 1'
+									) + (queryObj.queryPost || ''),
+									') OR ',
+								]
+								: []
+							),
+							'(',
+							queryObj.query.replace(/#$/, '*#') + (queryObj.queryPost || ''),
+							') OR (',
+							queryObj.query.replaceAll(' IS ', ' HAS ') + (queryObj.queryPost || ''),
+							')'
+						].filter(Boolean)
+					};
+				}
+			}
+			return menu.isSeparator(queryObj)
+				? queryObj
+				: void (0);
+		}).filter(Boolean).forEach((queryObj) => {
+			// Add separators
+			if (menu.isSeparator(queryObj)) {
+				if (!menu.isSeparator(menu.getLastEntry() || { entryText: '' })) {
+					menu.newSeparator(partialMenu);
+				}
+			} else {
+				// Create names for all entries
+				queryObj.name = queryObj.name.length > 40 ? queryObj.name.substring(0, 40) + ' ...' : queryObj.name;
+				// Entries
+				menu.newEntry({
+					menuName: partialMenu, entryText: queryObj.name, func: () => {
+						let query = queryObj.query;
+						if (query.includes('#') && !fb.GetFocusItem(true)) { fb.ShowPopupMessage('Can not evaluate query without a selection:\n' + queryObj.query, 'Quicksearch'); return; }
 						const bShift = utils.IsKeyPressed(VK_SHIFT);
 						const bCtrl = utils.IsKeyPressed(VK_CONTROL);
 						if (this.buttonsProperties.bEvalSel[1]) {